+++ edgy/conf/global_settings.py
@@ -35,16 +35,4 @@ class EdgySettings(MediaSettings):
     ptpython_config_file: str = "~/.config/ptpython/config.py"
 
     # General settings
-    filter_operators: Dict[str, str] = {
-        "exact": "__eq__",
-        "iexact": "ilike",
-        "contains": "like",
-        "icontains": "ilike",
-        "in": "in_",
-        "gt": "__gt__",
-        "gte": "__ge__",
-        "lt": "__lt__",
-        "lte": "__le__",
-    }
-    many_to_many_relation: str = "relation_{key}"
     dialects: Dict[str, str] = {"postgres": "postgres", "postgresql": "postgresql"}